+++ kwcoco/coco_dataset.py
@@ -27,6 +27,8 @@ An informal spec is as follows:
         'width': int  # the base width of this video (all associated images must have this width)
         'height': int  # the base height of this video (all associated images must have this height)
 
+        'resolution': int | str,  # indicates the size of a pixel in video space
+
         # In the future this may be extended to allow pointing to video files
     }
 
@@ -49,6 +51,7 @@ An informal spec is as follows:
         'height': int,  # pixel height of "base" image
 
         'channels': <ChannelSpec>,   # a string encoding of the channels in the main image (optional if auxiliary items are specified)
+        'resolution': int | str,  # indicates the size of a pixel in image space
 
         'auxiliary': [  # information about any auxiliary channels / bands
             {
@@ -62,7 +65,7 @@ An informal spec is as follows:
         ]
 
         'video_id': str  # if this image is a frame in a video sequence, this id is shared by all frames in that sequence.
-        'timestamp': str | int  # a iso-string timestamp or an integer in flicks.
+        'timestamp': str | int  # a iso-8601 or unix timestamp.
         'frame_index': int  # ordinal frame index which can be used if timestamp is unknown.
         'warp_img_to_vid': <TransformSpec>  # a transform image space to video space (identity if unspecified), can be used for sensor alignment or video stabilization
     }
@@ -243,8 +246,8 @@ An informal spec is as follows:
                    {"channels": "cloudmask": "file_name": "cloudmask.tif", "warp_aux_to_img": {"scale": 4.0}, "height": 100, "width": 100, ...},
                    {"channels": "nir": "file_name": "nir.tif", "warp_aux_to_img": {"scale": 2.0}, "height": 200, "width": 200, ...},
                    {"channels": "swir": "file_name": "swir.tif", "warp_aux_to_img": {"scale": 2.0}, "height": 200, "width": 200, ...},
-                   {"channels": "model1_predictions:0.6": "file_name": "model1_preds.tif", "warp_aux_to_img": {"scale": 8.0}, "height": 50, "width": 50, ...},
-                   {"channels": "model2_predictions:0.3": "file_name": "model2_preds.tif", "warp_aux_to_img": {"scale": 8.0}, "height": 50, "width": 50, ...},
+                   {"channels": "model1_predictions.0:6": "file_name": "model1_preds.tif", "warp_aux_to_img": {"scale": 8.0}, "height": 50, "width": 50, ...},
+                   {"channels": "model2_predictions.0:3": "file_name": "model2_preds.tif", "warp_aux_to_img": {"scale": 8.0}, "height": 50, "width": 50, ...},
                 ]
             }
 
@@ -272,7 +275,7 @@ An informal spec is as follows:
 
         {
             'video_id': str  # optional, if this image is a frame in a video sequence, this id is shared by all frames in that sequence.
-            'timestamp': int  # optional, timestamp (ideally in flicks), used to identify the timestamp of the frame. Only applicable video inputs.
+            'timestamp': str | int  # optional, an iso8601 or unix timestamp
             'frame_index': int  # optional, ordinal frame index which can be used if timestamp is unknown.
         }
 
@@ -291,9 +294,10 @@ Note:
 
 
 TODO:
-    - [ ] Use ijson to lazilly load pieces of the dataset in the background or
-      on demand. This will give us faster access to categories / images,
-      whereas we will always have to wait for annotations etc...
+    - [ ] Use ijson (modified to support NaN) to lazilly load pieces of the
+        dataset in the background or on demand. This will give us faster access
+        to categories / images, whereas we will always have to wait for
+        annotations etc...
 
     - [X] Should img_root be changed to bundle_dpath?
 
@@ -1684,6 +1688,42 @@ class MixinCocoExtras(object):
                 hashid_sidecar_fpath.write_text(json_w.dumps(hashid_cache_data))
         return self.hashid
 
+    @classmethod
+    def _cached_hashid_for(cls, fpath):
+        """
+        Lookup the cached hashid for a kwcoco json file if it exists.
+        """
+        coco_fpath = ub.Path(fpath)
+        enable_cache = coco_fpath.exists()
+
+        if enable_cache:
+            cache_dpath = (coco_fpath.parent / '_cache')
+            cache_fname = coco_fpath.name + '.hashid.cache'
+            hashid_sidecar_fpath = cache_dpath / cache_fname
+            # Generate current lookup key
+            fpath_stat = coco_fpath.stat()
+            status_key = {
+                'st_size': fpath_stat.st_size,
+                'st_mtime': fpath_stat.st_mtime
+            }
+            if hashid_sidecar_fpath.exists():
+                cached_data = json_r.loads(hashid_sidecar_fpath.read_text())
+                if cached_data['status_key'] == status_key:
+                    cached_data['hashid']
+                    cached_data['hashid_parts']
+                    cache_miss = False
+
+        if cache_miss:
+            self._build_hashid()
+            if enable_cache:
+                hashid_cache_data = {
+                    'hashid': self.hashid,
+                    'hashid_parts': self.hashid_parts,
+                    'status_key': status_key,
+                }
+                hashid_sidecar_fpath.parent.ensuredir()
+                hashid_sidecar_fpath.write_text(json_w.dumps(hashid_cache_data))
+
     def _dataset_id(self):
         """
         A human interpretable name that can be used to uniquely identify the
@@ -2331,19 +2371,23 @@ class MixinCocoObjects(object):
     This is an alternative vectorized ORM-like interface to the coco dataset
     """
 
-    def annots(self, aids=None, gid=None, trackid=None):
+    def annots(self, annot_ids=None, image_id=None, trackid=None, aids=None, gid=None):
         """
         Return vectorized annotation objects
 
         Args:
-            aids (List[int]): annotation ids to reference, if unspecified
-                 all annotations are returned.
+            annot_ids (List[int]):
+                annotation ids to reference, if unspecified all annotations are
+                returned. An alias is "aids", which may be removed in the future.
 
-            gid (int): return all annotations that belong to this image id.
-                 mutually exclusive with other arguments.
+            image_id (int):
+                return all annotations that belong to this image id.
+                Mutually exclusive with other arguments.
+                An alias is "gids", which may be removed in the future.
 
-            trackid (int): return all annotations that belong to this track.
-                 mutually exclusive with other arguments.
+            trackid (int):
+                return all annotations that belong to this track.
+                mutually exclusive with other arguments.
 
         Returns:
             kwcoco.coco_objects1d.Annots: vectorized annotation object
@@ -2364,27 +2408,32 @@ class MixinCocoObjects(object):
                 None,
             ]
         """
-        if gid is not None:
-            aids = sorted(self.index.gid_to_aids[gid])
+        if image_id is None:
+            image_id = gid
+        if annot_ids is None:
+            annot_ids = aids
+
+        if image_id is not None:
+            annot_ids = sorted(self.index.gid_to_aids[image_id])
 
         if trackid is not None:
-            aids = self.index.trackid_to_aids[trackid]
+            annot_ids = self.index.trackid_to_aids[trackid]
 
-        if aids is None:
-            aids = sorted(self.index.anns.keys())
+        if annot_ids is None:
+            annot_ids = sorted(self.index.anns.keys())
 
-        return Annots(aids, self)
+        return Annots(annot_ids, self)
 
-    def images(self, gids=None, video_id=None, names=None, vidid=None):
+    def images(self, image_ids=None, video_id=None, names=None, gids=None, vidid=None):
         """
         Return vectorized image objects
 
         Args:
-            gids (List[int]): image ids to reference, if unspecified
-                 all images are returned.
+            image_ids (List[int]): image ids to reference, if unspecified
+                 all images are returned. An alias is `gids`.
 
             video_id (int): returns all images that belong to this video id.
-                mutually exclusive with `gids` arg.
+                mutually exclusive with `image_ids` arg.
 
             names (List[str]):
                 lookup images by their names.
@@ -2406,6 +2455,9 @@ class MixinCocoObjects(object):
             >>> print(images)
             <Images(num=2)>
         """
+        if image_ids is None:
+            image_ids = gids
+
         if vidid is not None:
             ub.schedule_deprecation(
                 'kwcoco', 'vidid', 'argument of CocoDataset.images',
@@ -2415,23 +2467,24 @@ class MixinCocoObjects(object):
             video_id = vidid
 
         if video_id is not None:
-            gids = self.index.vidid_to_gids[video_id]
+            image_ids = self.index.vidid_to_gids[video_id]
 
         if names is not None:
-            gids = [self.index.name_to_img[name]['id'] for name in names]
+            image_ids = [self.index.name_to_img[name]['id'] for name in names]
 
-        if gids is None:
-            gids = sorted(self.index.imgs.keys())
+        if image_ids is None:
+            image_ids = sorted(self.index.imgs.keys())
 
-        return Images(gids, self)
+        return Images(image_ids, self)
 
-    def categories(self, cids=None):
+    def categories(self, category_ids=None, cids=None):
         """
         Return vectorized category objects
 
         Args:
-            cids (List[int]): category ids to reference, if unspecified
-                 all categories are returned.
+            category_ids (List[int]):
+                category ids to reference, if unspecified all categories are
+                returned. The `cids` argument is an alias.
 
         Returns:
             kwcoco.coco_objects1d.Categories: vectorized category object
@@ -2443,20 +2496,24 @@ class MixinCocoObjects(object):
             >>> print(categories)
             <Categories(num=8)>
         """
-        if cids is None:
-            cids = sorted(self.index.cats.keys())
-        return Categories(cids, self)
+        if category_ids is None:
+            category_ids = cids
+        if category_ids is None:
+            category_ids = sorted(self.index.cats.keys())
+        return Categories(category_ids, self)
 
-    def videos(self, vidids=None, names=None):
+    def videos(self, video_ids=None, names=None, vidids=None):
         """
         Return vectorized video objects
 
         Args:
-            vidids (List[int]): video ids to reference, if unspecified
-                 all videos are returned.
+            video_ids (List[int]): video ids to reference, if unspecified
+                 all videos are returned. The `vidids` argument is an alias.
+                 Mutually exclusive with other args.
 
             names (List[str]):
                 lookup videos by their name.
+                Mutually exclusive with other args.
 
         Returns:
             kwcoco.coco_objects1d.Videos: vectorized video object
@@ -2477,11 +2534,14 @@ class MixinCocoObjects(object):
             >>> videos.lookup('id')
             >>> print('videos.objs = {}'.format(ub.repr2(videos.objs[0:2], nl=1)))
         """
-        if vidids is None:
-            vidids = sorted(self.index.videos.keys())
+        if video_ids is None:
+            video_ids = vidids
+        if video_ids is None:
+            video_ids = sorted(self.index.videos.keys())
+
         if names is not None:
-            vidids = [self.index.name_to_video[name]['id'] for name in names]
-        return Videos(vidids, self)
+            video_ids = [self.index.name_to_video[name]['id'] for name in names]
+        return Videos(video_ids, self)
 
 
 class MixinCocoStats(object):
@@ -2550,7 +2610,10 @@ class MixinCocoStats(object):
 
                 ensure_imgsize (default=True): ensure image size is populated
 
-                legacy (default=False): if true tries to convert data
+                mmlab (default=False): if True tries to convert data
+                to be compatible with open-mmlab tooling.
+
+                legacy (default=False): if True tries to convert data
                 structures to items compatible with the original
                 pycocotools spec
 
@@ -2616,6 +2679,14 @@ class MixinCocoStats(object):
                                 '''))
                         else:
                             ann['area'] = w * h
+
+        if config.get('mmlab', False):
+            ### mmdet wants frame_id not frame_index, perhaps other properties
+            ### TODO: add more mmlab support
+            if self.dataset.get('videos', None):
+                for images in self.videos().images:
+                    for frame_index, img in enumerate(images.objs):
+                        img['frame_id'] = frame_index
 
         if config.get('legacy', False):
             try:
@@ -2725,8 +2796,6 @@ class MixinCocoStats(object):
             except jsonschema.exceptions.ValidationError as ex:
                 err = ex
                 print(f'err.absolute_path={err.absolute_path}')
-                import xdev
-                xdev.embed()
                 msg = 'Failed to validate schema: {}'.format(str(err))
                 _error(msg)